Implement readAWS for the mappers, and now 'populate' is much simpler
