+++ src/mapper/ami.ts
@@ -6,6 +6,7 @@ import { BootModeMapper, } from './boot_mode';
 import { CPUArchitectureMapper, } from './cpu_architecture';
 import { EBSBlockDeviceMappingMapper, } from './ebs_block_device_mapping';
 import { EntityMapper, } from './entity';
+import * as Entities from '../entity'; // TODO: Is there a way to avoid this double import?
 import { IndexedAWS, } from '../services/indexed-aws'
 import { ProductCodeMapper, } from './product_code';
 import { StateReasonMapper, } from './state_reason';
@@ -56,7 +57,40 @@ export const AMIMapper = new EntityMapper(AMI, {
     ami.Tags.map(tag => TagMapper.fromAWS(tag, indexes)) :
     [],
 }, {
-  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => { throw new Error('tbd') },
+  readAWS: async (awsClient: AWS, indexes: IndexedAWS) => {
+    const t1 = Date.now();
+    const amis = (await awsClient.getAMIs())?.Images ?? [];
+    indexes.setAll(AMI, amis, 'ImageId');
+    const t2 = Date.now();
+    console.log(`AMIs set in ${t2 - t1}ms`);
+    // Set aux AMI indexes, too
+    for (const ami of amis) {
+      if (ami.Architecture) {
+        indexes.set(Entities.CPUArchitecture, ami.Architecture, ami.Architecture);
+      }
+      if (ami.ProductCodes && ami.ProductCodes.length) {
+        for (const pc of ami.ProductCodes) {
+          if (pc.ProductCodeId) {
+            indexes.set(Entities.ProductCode, pc.ProductCodeId, pc);
+          } else {
+            throw Error('productCodes is this possible?');
+          }
+        }
+      }
+      if (ami.StateReason) {
+        if (ami.StateReason.Code) {
+          indexes.set(Entities.StateReason, ami.StateReason.Code, ami.StateReason);
+        } else {
+          throw Error('stateReason is this possible?')
+        }
+      }
+      if (ami.BootMode) {
+        indexes.set(Entities.BootMode, ami.BootMode, ami.BootMode);
+      }
+    }
+    const t3 = Date.now();
+    console.log(`AMI sub entities set in ${t3 - t2}ms`);
+  },
   createAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   updateAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   deleteAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
