+++ src/mapper/boot_mode.ts
@@ -8,7 +8,10 @@ import { IndexedAWS, } from '../services/indexed-aws'
 export const BootModeMapper = new EntityMapper(BootMode, {
   mode: (bootMode: BootModeValues, _indexes: IndexedAWS) => bootMode,
 }, {
-  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => { throw new Error('tbd') },
+  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => {
+    // Handled by AMI
+    return
+  },
   createAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   updateAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   deleteAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
