+++ src/mapper/cpu_architecture.ts
@@ -6,7 +6,10 @@ import { IndexedAWS, } from '../services/indexed-aws'
 export const CPUArchitectureMapper = new EntityMapper(CPUArchitecture, {
   cpuArchitecture: (cpuArchitecure: string, _indexes: IndexedAWS) => cpuArchitecure,
 }, {
-  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => { throw new Error('tbd') },
+  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => {
+    // Handled by AMI
+    return
+  },
   createAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   updateAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   deleteAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
