+++ src/mapper/product_code.ts
@@ -9,7 +9,10 @@ export const ProductCodeMapper = new EntityMapper(ProductCode, {
   productCodeId: (productCode: ProductCodeAWS, _indexes: IndexedAWS) => productCode?.ProductCodeId,
   productCodeType: (productCode: ProductCodeAWS, _indexes: IndexedAWS) => productCode?.ProductCodeType,
 }, {
-  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => { throw new Error('tbd') },
+  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => {
+    // Handled by AMI
+    return
+  },
   createAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   updateAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   deleteAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
