+++ src/mapper/region.ts
@@ -11,8 +11,11 @@ export const RegionMapper = new EntityMapper(Region, {
   optInStatus: (region: RegionAWS, _indexes: IndexedAWS) => region?.OptInStatus,
 }, {
   readAWS: async (awsClient: AWS, indexes: IndexedAWS) => {
+    const t1 = Date.now();
     const regions = (await awsClient.getRegions())?.Regions ?? [];
     indexes.setAll(Region, regions, 'RegionName');
+    const t2 = Date.now();
+    console.log(`Regions set in ${t2 - t1}ms`);
   },
   createAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   updateAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
