+++ src/mapper/state_reason.ts
@@ -9,7 +9,10 @@ export const StateReasonMapper = new EntityMapper(StateReason, {
   code: (sr: StateReasonAWS, _indexes: IndexedAWS) => sr?.Code,
   message: (sr: StateReasonAWS, _indexes: IndexedAWS) => sr?.Message,
 }, {
-  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => { throw new Error('tbd') },
+  readAWS: async (_awsClient: AWS, _indexes: IndexedAWS) => {
+    // Set by AMI
+    return
+  },
   createAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   updateAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
   deleteAWS: async (_obj: any, _indexes: IndexedAWS) => { throw new Error('tbd') },
